// types/auth.ts
export interface FormData {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  confirmPassword: string;
  skills: string[];
  interests: string[];
  professionalGoals: string[];
  values: string[];
  bio: string;
  meaningfulGoal: string;
  location: string;
  jobTitle: string;
  company: string;
  industry: string;
  yearsOfExperience: string;
  experience: string;
  education: string;
  linkedinUrl?: string;
  twitterUrl?: string;
  githubUrl?: string;
  websiteUrl?: string;
  avatarUrl?: string;
} 